pyproject.toml
[tool.poetry]
name = "website"
version = "0.1.0"
description = ""
authors = ["Marc Partensky <marc.partensky@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "*"
python-dotenv = "*"
channels = "*"
django-storages = "*"
django-crispy-forms = "*"
django-extensions = "*"
django-allauth = "*"
django-compressor = "*"
django-markdown-view = "*"
django-mdeditor = "*"
django-ftpserver = "*"
django-sql-explorer = "*"
django-q = "*"
django-robots = "*"
shortuuid = "*"
django-shortuuidfield = "*"
django-jsonfield-backport = "*"
rich = "*"
djangorestframework = "*"
PyYAML = "*"
django-cors-headers = "*"
Pillow = "*"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "*"
pylint = "*"
neovim = "*"
black = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
